==> notebooks/02_preprocess_merge.ipynb <==
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import numpy as np
from pathlib import Path
sns.set()

# 1. Ucitavanje podataka

ratings = pd.read_csv("../data/raw/ratings.csv")
movies = pd.read_csv("../data/raw/movies.csv")
tags = pd.read_csv("../data/raw/tags.csv")

# 2. Preprocesiranje

### Uklanjanje filmova bez žanra

movies = movies[movies['genres'] != '(no genres listed)'].copy()

### Dodavanje kolona rating_count i mean_rating

rating_stats = ratings.groupby('movieId').agg(
    mean_rating=('rating', 'mean'),
    rating_count=('rating', 'count')
).reset_index()

movies = pd.merge(movies, rating_stats, on='movieId', how='left')
movies['rating_count'] = movies['rating_count'].fillna(0).astype(int)
movies['mean_rating']  = movies['mean_rating'].astype(float)

### Čišćenje i grupisanje tagova zajedno sa naslovima filmova

def clean_tag(t: str) -> str:
    t = str(t).lower().strip()
    return " ".join(t.split())

if not tags.empty:
    tags['tag'] = tags['tag'].astype(str).apply(clean_tag)
    tags = tags[tags['tag'] != '']  # uklanjanje praznih
    tags_grouped = tags.groupby('movieId')['tag'].apply(lambda x: " ".join(x)).reset_index()
else:
    tags_grouped = pd.DataFrame({'movieId': [], 'tag': []})

movies = pd.merge(movies, tags_grouped, on='movieId', how='left')
movies['tag'] = movies['tag'].fillna('')

movies["title_clean"] = (
    movies["title"]
    .str.replace(r"\(\d{4}\)", "", regex=True)      # izbaci godinu
    .str.lower()
    .str.replace("[^a-z0-9 ]", " ", regex=True)     # samo alfanumerički karakteri
)

movies["text_features"] = movies["title_clean"] + " " + movies["tag"].fillna("")

### Kreiranje dummy (one-hot) kolona za žanrove

movies['genres_tok'] = movies['genres'].str.replace('|', ' ', regex=False).str.lower()

genre_dummies = movies['genres_tok'].str.get_dummies(sep=' ')

# Prefix
genre_dummies = genre_dummies.add_prefix('genre_')

# Spoji dummy kolone nazad sa glavnim DataFrame-om
movies = pd.concat([movies, genre_dummies], axis=1)

print(f"Kreirano {genre_dummies.shape[1]} žanrovskih dummy kolona")

### Dropovanje redundantnih kolona

movies = movies.drop(columns=['genres', 'genres_tok', 'tag', 'title_clean'], errors='ignore')

print(movies.shape)
movies.head()

# 3. Čuvanje nove film tabele

output_path = Path("../data/processed/movies_merged.parquet")
movies.to_parquet(output_path, index=False)
print("Sačuvano:", output_path)
